pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.13",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
]
